% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{read_sql}
\alias{read_sql}
\title{Read in SQL File to Dataframe}
\usage{
read_sql(FUN = con_ob(), file)
}
\arguments{
\item{FUN}{name of function used to connect to database, con_ob() or con_cs()}

\item{file}{sql file name to be read in}
}
\value{

}
\description{
Read in SQL File to Dataframe
}
