% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{read_sql_params}
\alias{read_sql_params}
\title{Read in SQL File with Parameters to Dataframe}
\usage{
read_sql_params(FUN = con_ob(), file, list)
}
\arguments{
\item{FUN}{name of function used to connect to database, con_ob() or con_cs()}

\item{file}{sql file name to be read in}

\item{list}{list of parameters that are in the SQL code, may need use of paste for list (list <- list(paste(param), paste(param)))}
}
\value{

}
\description{
Read in SQL File with Parameters to Dataframe
}
